requirements.txt
torch
transformers
Pillow
streamlit
PyMuPDF
Pillow
gtts
pyttsx3
pygame
streamlit
opencv-python==4.5.4.58
numpy